+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.55](https://github.com/realityforge/proton/tree/v0.55) (2022-04-22) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.54...v0.55)
 
 Changes in this release: