+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.56](https://github.com/realityforge/proton/tree/v0.56) (2022-04-27) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.55...v0.56)
+
+Changes in this release:
 
 * Ensure `SynthesizingProcessor` supports source in version `17`.
 * Introduce `CompileTestUtil` and `Compilation` classes in the `qa` module to reimplement functionality present in the `com.google.testing.compile:compile-testing` artifact. Rewrite `AbstractProcessorTest` to use `CompileTestUtil` functionality rather than the equivalent code in the `com.google.testing.compile:compile-testing` artifact. Remove the `com.google.testing.compile:compile-testing` artifact and related dependencies. This resulted in simpler code and made it easier to evolve the project to support the latest version of Java.